+++ tests/modules/test_lint.py
@@ -210,14 +210,6 @@ class TestModulesCreate(TestModules):
         assert len(module_lint.passed) > 0
         assert len(module_lint.warned) >= 0
 
-    def test_modules_lint_update_meta_yml(self):
-        """update the meta.yml of a module"""
-        module_lint = nf_core.modules.ModuleLint(directory=self.nfcore_modules, fix=True)
-        module_lint.lint(print_results=False, module="fastqc")
-        assert len(module_lint.failed) == 0, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
-        assert len(module_lint.passed) > 0
-        assert len(module_lint.warned) >= 0
-
     def test_modules_lint_no_gitlab(self):
         """Test linting a pipeline with no modules installed"""
         self.mods_remove.remove("fastqc", force=True)
@@ -282,7 +274,7 @@ class TestModulesCreate(TestModules):
                 all_modules=True,
             )
 
-        assert len(module_lint.failed) == 1
+        assert len(module_lint.failed) == 1, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
         assert len(module_lint.passed) > 0
         assert len(module_lint.warned) >= 0
 
@@ -312,6 +304,14 @@ class TestModulesCreate(TestModules):
             assert (
                 len(mocked_ModuleLint.failed) == failed
             ), f"{test}: Expected {failed} FAIL, got {len(mocked_ModuleLint.failed)}."
+
+    def test_modules_lint_update_meta_yml(self):
+        """update the meta.yml of a module"""
+        module_lint = nf_core.modules.lint.ModuleLint(directory=self.nfcore_modules, fix=True)
+        module_lint.lint(print_results=False, module="bpipe/test")
+        assert len(module_lint.failed) == 0, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
+        assert len(module_lint.passed) > 0
+        assert len(module_lint.warned) >= 0
 
     def test_modules_lint_snapshot_file(self):
         """Test linting a module with a snapshot file"""
@@ -521,25 +521,6 @@ class TestModulesCreate(TestModules):
         assert len(module_lint.warned) >= 0
         assert module_lint.failed[0].lint_test == "meta_yml_valid"
 
-    def test_modules_meta_yml_input_mismatch(self):
-        """Test linting a module with an extra entry in input fields in meta.yml compared to module.input"""
-        with open(Path(self.nfcore_modules, "modules", "nf-core", "bpipe", "test", "main.nf")) as fh:
-            main_nf = fh.read()
-        main_nf_new = main_nf.replace("path bam", "path bai")
-        with open(Path(self.nfcore_modules, "modules", "nf-core", "bpipe", "test", "main.nf"), "w") as fh:
-            fh.write(main_nf_new)
-        module_lint = nf_core.modules.lint.ModuleLint(directory=self.nfcore_modules)
-        module_lint.lint(print_results=False, module="bpipe/test")
-        with open(Path(self.nfcore_modules, "modules", "nf-core", "bpipe", "test", "main.nf"), "w") as fh:
-            fh.write(main_nf)
-        assert len(module_lint.failed) == 0, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
-        assert len(module_lint.passed) >= 0
-        assert len(module_lint.warned) == 2, f"Linting warning with {[x.__dict__ for x in module_lint.warned]}"
-        lint_tests = [x.lint_test for x in module_lint.warned]
-        # check that it is there twice:
-        assert lint_tests.count("meta_input_meta_only") == 1
-        assert lint_tests.count("meta_input_main_only") == 1
-
     def test_modules_meta_yml_output_mismatch(self):
         """Test linting a module with an extra entry in output fields in meta.yml compared to module.output"""
         with open(Path(self.nfcore_modules, "modules", "nf-core", "bpipe", "test", "main.nf")) as fh:
@@ -551,13 +532,9 @@ class TestModulesCreate(TestModules):
         module_lint.lint(print_results=False, module="bpipe/test")
         with open(Path(self.nfcore_modules, "modules", "nf-core", "bpipe", "test", "main.nf"), "w") as fh:
             fh.write(main_nf)
-        assert len(module_lint.failed) == 0, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
+        assert len(module_lint.failed) == 1, f"Linting failed with {[x.__dict__ for x in module_lint.failed]}"
         assert len(module_lint.passed) >= 0
-        assert len(module_lint.warned) == 2
-        lint_tests = [x.lint_test for x in module_lint.warned]
-        # check that it is there twice:
-        assert lint_tests.count("meta_output_meta_only") == 1
-        assert lint_tests.count("meta_output_main_only") == 1
+        assert "Module `meta.yml` does not match `main.nf`" in module_lint.failed[0].message
 
     def test_modules_meta_yml_incorrect_name(self):
         """Test linting a module with an incorrect name in meta.yml"""